reorg, stuff

+++ src/app/app/tags/tagActions.ts
@@ -1,14 +1,21 @@
 "use server";
 
+import { getUser, getUserId } from "@/app/login/action";
 import { db } from "@/lib/db/db";
 import { tags } from "@/lib/db/schema";
 import { createId, returnErr, returnSuccess } from "@/lib/utils";
-import { desc, eq } from "drizzle-orm";
+import { and, desc, eq } from "drizzle-orm";
 import { revalidateTag } from "next/cache";
 import { is, minLength, object, string } from "valibot";
 
 export async function getTags() {
-	return db.query.tags.findMany({ orderBy: [desc(tags.createdAt)] });
+	const userId = await getUserId(true);
+
+	return db.query.tags.findMany({
+		orderBy: [desc(tags.createdAt)],
+		where: eq(tags.userId, userId),
+		limit: 50,
+	});
 }
 
 const createTagSchema = object({
@@ -18,6 +25,8 @@ const createTagSchema = object({
 
 export async function createTag(formData: FormData) {
 	const data = Object.fromEntries(formData.entries());
+
+	const user = await getUser(true);
 
 	if (!is(createTagSchema, data)) {
 		return returnErr("invalid data");
@@ -29,7 +38,7 @@ export async function createTag(formData: FormData) {
 			name: data.name,
 			color: data.color,
 			createdAt: new Date(),
-			userId: "1",
+			userId: user.id,
 		});
 	} catch (err) {
 		console.error("failed to insert tag to db", err);
@@ -39,7 +48,7 @@ export async function createTag(formData: FormData) {
 
 	revalidateTag("/tags");
 
-	return returnSuccess(null);
+	return returnSuccess();
 }
 
 const editTagSchema = object({
@@ -55,6 +64,8 @@ export async function editTag(formData: FormData) {
 		return returnErr("invalid data");
 	}
 
+	const userId = await getUserId(true);
+
 	try {
 		await db
 			.update(tags)
@@ -62,7 +73,7 @@ export async function editTag(formData: FormData) {
 				name: data.name,
 				color: data.color,
 			})
-			.where(eq(tags.id, data.tagId));
+			.where(and(eq(tags.id, data.tagId), eq(tags.userId, userId)));
 	} catch (err) {
 		console.error("failed to update tag in db", err);
 
@@ -71,12 +82,14 @@ export async function editTag(formData: FormData) {
 
 	revalidateTag("/tags");
 
-	return returnSuccess(null);
+	return returnSuccess();
 }
 
 export async function deleteTag(tagId: string) {
+	const userId = await getUserId(true);
+
 	try {
-		await db.delete(tags).where(eq(tags.id, tagId));
+		await db.delete(tags).where(and(eq(tags.id, tagId), eq(tags.userId, userId)));
 	} catch (err) {
 		console.error(`failed to delete tag (tagId: '${tagId}') from db `, err);
 
@@ -85,5 +98,5 @@ export async function deleteTag(tagId: string) {
 
 	revalidateTag("/tags");
 
-	return returnSuccess(null);
+	return returnSuccess();
 }
